requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
gTTS==2.5.0
textblob==0.18.0
nltk==3.8.1
